Extract PlayerInfoDialog out of PaperdollDialog
