+++ EndlessClient/Dialogs/PaperdollDialog.cs
@@ -4,23 +4,17 @@ using EndlessClient.ControlSets;
 using EndlessClient.Dialogs.Extensions;
 using EndlessClient.Dialogs.Factories;
 using EndlessClient.Dialogs.Services;
-using EndlessClient.GameExecution;
 using EndlessClient.HUD;
 using EndlessClient.HUD.Controls;
 using EndlessClient.HUD.Inventory;
 using EndlessClient.HUD.Panels;
-using EOLib;
 using EOLib.Domain.Character;
-using EOLib.Domain.Online;
-using EOLib.Extensions;
 using EOLib.Graphics;
 using EOLib.IO;
 using EOLib.IO.Repositories;
 using EOLib.Localization;
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Optional;
-using Optional.Unsafe;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -28,10 +22,8 @@ using XNAControls;
 
 namespace EndlessClient.Dialogs
 {
-    public class PaperdollDialog : BaseEODialog
+    public class PaperdollDialog : PlayerInfoDialog<PaperdollData>
     {
-        private static readonly Rectangle _iconDrawRect = new Rectangle(227, 258, 44, 21);
-
         private readonly IInventoryController _inventoryController;
         private readonly IPaperdollProvider _paperdollProvider;
         private readonly IPubFileProvider _pubFileProvider;
@@ -40,27 +32,14 @@ namespace EndlessClient.Dialogs
         private readonly IEOMessageBoxFactory _eoMessageBoxFactory;
         private readonly IStatusLabelSetter _statusLabelSetter;
         private readonly ISfxPlayer _sfxPlayer;
-        private readonly bool _isMainCharacter;
-        private readonly Texture2D _characterIconSheet;
-        private Option<Rectangle> _characterIconSourceRect;
+
         private readonly InventoryPanel _inventoryPanel;
 
         private Option<PaperdollData> _paperdollData;
 
         private readonly List<PaperdollDialogItem> _childItems;
 
-        private readonly IXNALabel _name,
-            _home,
-            _class,
-            _partner,
-            _title,
-            _guild,
-            _rank;
-
-        public Character Character { get; }
-
-        public PaperdollDialog(IGameStateProvider gameStateProvider,
-                               INativeGraphicsManager nativeGraphicsManager,
+        public PaperdollDialog(INativeGraphicsManager nativeGraphicsManager,
                                IInventoryController inventoryController,
                                IPaperdollProvider paperdollProvider,
                                IPubFileProvider pubFileProvider,
@@ -71,7 +50,7 @@ namespace EndlessClient.Dialogs
                                IStatusLabelSetter statusLabelSetter,
                                ISfxPlayer sfxPlayer,
                                Character character, bool isMainCharacter)
-            : base(nativeGraphicsManager, gameStateProvider)
+            : base(nativeGraphicsManager, eoDialogButtonService, pubFileProvider, character, isMainCharacter)
         {
             _paperdollProvider = paperdollProvider;
             _pubFileProvider = pubFileProvider;
@@ -81,10 +60,6 @@ namespace EndlessClient.Dialogs
             _statusLabelSetter = statusLabelSetter;
             _sfxPlayer = sfxPlayer;
             _inventoryController = inventoryController;
-            Character = character;
-            _isMainCharacter = isMainCharacter;
-            _characterIconSheet = GraphicsManager.TextureFromResource(GFXTypes.PostLoginUI, 32, true);
-            _characterIconSourceRect = Option.None<Rectangle>();
 
             _inventoryPanel = _hudControlProvider.GetComponent<InventoryPanel>(HudControlIdentifier.InventoryPanel);
 
@@ -92,42 +67,6 @@ namespace EndlessClient.Dialogs
 
             BackgroundTexture = GraphicsManager.TextureFromResource(GFXTypes.PostLoginUI, 49);
             BackgroundTextureSource = new Rectangle(0, BackgroundTexture.Height / 2 * Character.RenderProperties.Gender, DrawArea.Width, BackgroundTexture.Height / 2);
-
-            var okButton = new XNAButton(eoDialogButtonService.SmallButtonSheet,
-                new Vector2(276, 253),
-                eoDialogButtonService.GetSmallDialogButtonOutSource(SmallButton.Ok),
-                eoDialogButtonService.GetSmallDialogButtonOverSource(SmallButton.Ok));
-            okButton.OnClick += (_, _) => Close(XNADialogResult.OK);
-            okButton.Initialize();
-            okButton.SetParentControl(this);
-
-            _name = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 22, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _name.Initialize();
-            _name.SetParentControl(this);
-
-            _home = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 52, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _home.Initialize();
-            _home.SetParentControl(this);
-
-            _class = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 82, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _class.Initialize();
-            _class.SetParentControl(this);
-
-            _partner = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 112, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _partner.Initialize();
-            _partner.SetParentControl(this);
-
-            _title = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 142, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _title.Initialize();
-            _title.SetParentControl(this);
-
-            _guild = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 202, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _guild.Initialize();
-            _guild.SetParentControl(this);
-
-            _rank = new XNALabel(Constants.FontSize08pt5) { DrawArea = new Rectangle(228, 232, 1, 1), ForeColor = ColorConstants.LightGrayText };
-            _rank.Initialize();
-            _rank.SetParentControl(this);
 
             CenterInGameView();
 
@@ -158,37 +97,9 @@ namespace EndlessClient.Dialogs
             base.OnUpdateControl(gameTime);
         }
 
-        protected override void OnDrawControl(GameTime gameTime)
+        protected override void UpdateDisplayedData(PaperdollData paperdollData)
         {
-            base.OnDrawControl(gameTime);
-
-            _spriteBatch.Begin();
-
-            _characterIconSourceRect.MatchSome(sourceRect =>
-            {
-                _spriteBatch.Draw(_characterIconSheet,
-                    new Vector2(
-                        DrawAreaWithParentOffset.X + _iconDrawRect.X + (_iconDrawRect.Width / 2) - (sourceRect.Width / 2),
-                        DrawAreaWithParentOffset.Y + _iconDrawRect.Y + (_iconDrawRect.Height / 2) - (sourceRect.Height / 2)),
-                    sourceRect,
-                    Color.White);
-            });
-
-            _spriteBatch.End();
-        }
-
-        private void UpdateDisplayedData(PaperdollData paperdollData)
-        {
-            _name.Text = Capitalize(paperdollData.Name);
-            _home.Text = Capitalize(paperdollData.Home);
-
-            paperdollData.Class.SomeWhen(x => x != 0)
-                .MatchSome(classId => _class.Text = Capitalize(_pubFileProvider.ECFFile[classId].Name));
-
-            _partner.Text = Capitalize(paperdollData.Partner);
-            _title.Text = Capitalize(paperdollData.Title);
-            _guild.Text = Capitalize(paperdollData.Guild);
-            _rank.Text = Capitalize(paperdollData.Rank);
+            base.UpdateDisplayedData(paperdollData);
 
             foreach (var control in _childItems)
             {
@@ -264,17 +175,6 @@ namespace EndlessClient.Dialogs
 
                 _childItems.Add(paperdollItem);
             }
-
-            _characterIconSourceRect = Option.Some(GetOnlineIconSourceRectangle(paperdollData.Icon));
-        }
-
-        private static string Capitalize(string input) => 
-            string.IsNullOrEmpty(input) ? string.Empty : char.ToUpper(input[0]) + input[1..].ToLower();
-
-        private static Rectangle GetOnlineIconSourceRectangle(OnlineIcon icon)
-        {
-            var (x, y, width, height) = icon.ToChatIcon().GetChatIconRectangleBounds().ValueOrDefault();
-            return new Rectangle(x, y, width, height);
         }
     }
 }
